backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
markdown
python-decouple
mysqlclient